Tidy Settings before adding new option.
